pyproject.toml
[tool.poetry]
name = "plant-id"
version = "0.1.0"
description = ""
authors = ["Justin Bennington <justin@somewhere.systems>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.1"
pillow = "^10.4.0"
datasets = "^2.20.0"
torchvision = "^0.18.1"
ftfy = "^6.0.3"
regex = "^2021.8.28"
tqdm = "^4.62.2"
openai-clip = "^1.0.1"
coremltools = "^7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
